+++ ring_buf.h
@@ -4,7 +4,7 @@
 
 typedef struct{
     float value;
-    uint32_t timestamp_ms;
+    long long timestamp_ms;
 } Entry;
 
 typedef struct{
@@ -17,7 +17,7 @@ typedef struct{
 
 RingBuf rb_init(size_t capacity);
 
-bool rb_push(RingBuf *rb, float value, uint32_t timestamp_ms);
+bool rb_push(RingBuf *rb, float value, long long timestamp_ms);
 bool rb_pop(RingBuf *rb, Entry *out);
 bool rb_peek(const RingBuf *rb, Entry *out);
 size_t rb_count(const RingBuf *rb);
@@ -25,4 +25,4 @@ bool rb_full(const RingBuf *rb);
 void rb_free(RingBuf *rb);
 float rb_avg(RingBuf *rb);
 
-long millis();
+long long millis();
